# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FactSnap API",
        "description": "API for managing educational categories, packages, and facts.\n *\n * # Authentication\n * \n * This API supports **dual authentication**:\n * \n * ## Firebase Authentication (for regular users)\n * - **Token Type**: Firebase ID Token\n * - **Format**: `Authorization: Bearer <firebase_id_token>`\n * - **Access**: 'free', 'loggedIn', and 'premium' packages (if premium user)\n * \n * ## Sanctum Authentication (for guest users)  \n * - **Token Type**: Sanctum Bearer Token (10-year expiration)\n * - **Format**: `Authorization: Bearer <sanctum_token>`\n * - **Access**: 'free' and 'loggedIn' packages ('premium' if created as premium guest)\n * \n * ## Package Access Levels\n * - **free**: No authentication required\n * - **loggedIn**: Requires authentication (Firebase OR Sanctum)\n * - **premium**: Requires premium user status",
        "contact": {
            "email": "support@factsnap.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Factsnap API Server"
        }
    ],
    "paths": {
...
        "/packages/{id}": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "Get a specific package with its category and facts",
                "description": "Display a specific package with its category and facts.",
                "operationId": "ae4e1772df8d6e55efccd5495f46921e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Package UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Package retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/Package"
                                                },
                                                {
                                                    "properties": {
                                                        "category": {
                                                            "$ref": "#/components/schemas/Category"
                                                        },
                                                        "facts": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/Fact"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Package not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/packages": {
            "get": {
                "tags": [
                    "User Packages"
                ],
                "summary": "Get user's subscribed packages",
                "description": "Retrieve all packages that the authenticated user has subscribed to. Supports both Firebase authentication (ID tokens) and Guest authentication (Sanctum tokens).",
                "operationId": "getUserPackages",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "icon_url": {
                                                        "type": "string"
                                                    },
                                                    "access_level": {
                                                        "type": "string"
                                                    },
                                                    "subscribed_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "category": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "format": "uuid"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing Firebase ID token or Guest Sanctum token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "dualAuth": []
                    },
                    {
                        "firebaseAuth": []
                    },
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/user/packages/{packageId}/subscribe": {
            "post": {
                "tags": [
                    "User Packages"
                ],
                "summary": "Subscribe to a package",
                "description": "Subscribe the authenticated user to a specific package. Supports both Firebase authentication (ID tokens) and Guest authentication (Sanctum tokens). Access levels: 'free' (all users), 'loggedIn' (Firebase users and guests), 'premium' (premium users only).",
                "operationId": "subscribeToPackage",
                "parameters": [
                    {
                        "name": "packageId",
                        "in": "path",
                        "description": "UUID of the package to subscribe to",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully subscribed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully subscribed to package: Package Name"
                                        },
                                        "data": {
                                            "properties": {
                                                "package_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "package_name": {
                                                    "type": "string"
                                                },
                                                "subscribed_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Already subscribed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied - user access level insufficient for package",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Package not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing Firebase ID token or Guest Sanctum token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "dualAuth": []
                    },
                    {
                        "firebaseAuth": []
                    },
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/user/packages/{packageId}/unsubscribe": {
            "delete": {
                "tags": [
                    "User Packages"
                ],
                "summary": "Unsubscribe from a package",
                "description": "Unsubscribe the authenticated user from a specific package. Supports both Firebase authentication (ID tokens) and Guest authentication (Sanctum tokens).",
                "operationId": "unsubscribeFromPackage",
                "parameters": [
                    {
                        "name": "packageId",
                        "in": "path",
                        "description": "UUID of the package to unsubscribe from",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully unsubscribed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully unsubscribed from package: Package Name"
                                        },
                                        "data": {
                                            "properties": {
                                                "package_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "package_name": {
                                                    "type": "string"
                                                },
                                                "unsubscribed_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Not subscribed to this package",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Package not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing Firebase ID token or Guest Sanctum token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "dualAuth": []
                    },
                    {
                        "firebaseAuth": []
                    },
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "properties": {
                    "id": {
                        "description": "Category UUID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "Category name",
                        "type": "string"
                    },
                    "icon_url": {
                        "description": "Category icon URL",
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Package": {
                "properties": {
                    "id": {
                        "description": "Package UUID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "category_id": {
                        "description": "Category UUID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "Package name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Package description",
                        "type": "string"
                    },
                    "icon_url": {
                        "description": "Package icon URL",
                        "type": "string"
                    },
                    "access_level": {
                        "description": "Access level",
                        "type": "string",
                        "enum": [
                            "free",
                            "premium"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "firebaseAuth": {
                "type": "http",
                "description": "Firebase ID Token obtained from Firebase SDK. Use in Authorization header: Bearer <firebase_id_token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctumAuth": {
                "type": "http",
                "description": "Sanctum Bearer Token for guest users. Get token from POST /api/auth/guest. Use in Authorization header: Bearer <sanctum_token>",
                "bearerFormat": "Token",
                "scheme": "bearer"
            },
            "dualAuth": {
                "type": "http",
                "description": "Accepts either Firebase ID tokens OR Sanctum tokens. Firebase is tried first, then Sanctum as fallback.",
                "scheme": "bearer"
            }
        }
    },
...
}